feat: add steps easing & clamp function & noise demo
