+++ demo/src/sandboxes/noise/.prettierrc
@@ -1,0 +1,9 @@
+{
+  "arrowParens": "avoid",
+  "jsxBracketSameLine": true,
+  "printWidth": 120,
+  "semi": false,
+  "singleQuote": true,
+  "tabWidth": 2,
+  "trailingComma": "es5"
+}
