+++ demo/src/sandboxes/noise/src/index.tsx
@@ -1,0 +1,12 @@
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import "./index.css";
+
+import App from "./App";
+
+const rootElement = document.getElementById("root")!;
+createRoot(rootElement).render(
+  <StrictMode>
+    <App />
+  </StrictMode>
+);
