+++ packages/shared/src/clamp.ts
@@ -1,0 +1,2 @@
+export const clamp = (min: number, max: number, v: number) =>
+  Math.min(Math.max(v, min), max)
